+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 ### Fixed
 
 - [TD-5751] Enrich domains parents in V2
+- [TD-5733] Enrich children's alias in V2
 
 ## [5.6.0] 2023-05-09